requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-functions==1.*
